requirements.txt
Django>=4.2,<5
uWSGI
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp-proto-http
dotenv